Basic User Interface changes
No need to take in these changes. I just made them as I am training on
the code. They do reflect ideas I have on the UI and that I think we
should adopt, but I am not sure I made the changes in the right
templates, as some templates may be core telescope and not sb-templates…