Anonymizing
I anonymized notifications where we could know who posted an idea, so
other users cannot figure out who posted an idea with notifications. It
would be great if we could also add in these cases the channel in which
the idea was posted - I just don’t know which variable we have for
that.